+++ composer.json
@@ -12,7 +12,8 @@
         "laravel/sanctum": "^3.2",
         "laravel/tinker": "^2.8",
         "laravel/ui": "^4.2",
-        "livewire/livewire": "^2.12"
+        "livewire/livewire": "^2.12",
+        "rafaelwendel/phpsupabase": "^0.0.7"
     },
     "require-dev": {
         "fakerphp/faker": "^1.9.1",
