+++ config/supabase.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'url' => env('SUPABASE_URL'),
+    'api_key' => env('SUPABASE_API_KEY'),
+];
